Cleanup version dependent configs, +others
It turns out use-package's :if, :unless, and :when only prevent
requiring a package, but not preventing installation. Magit and Paradox,
for instance, both pitch a fit if Emacs is too old.
Additionally, :disabled is a flag that does not take any arguments.

Lastly, put elisp-slime-nav-mode hook in the right place.